// src/ContentDashboard/Client/src/api/types.gen.ts
export type ContentOwnerResponseModel = {
    id: string;
    name: string;
    email?: null | string;
    documentCount: number;
};

export type DocumentOwnerModel = {
    id?: null | string;
    name: string;
};
// ...
export type ScheduleAction = 'Publish' | 'Unpublish';

export type ScheduledDocumentResponseModel = {
    id: string;
    name: string;
    action: ScheduleAction;
    date: string;
    culture?: null | string;
    owner?: null | DocumentOwnerModel;
};